Day 9 part 1

+++ src/main/java/intcode/ReadParameter.java
@@ -1,6 +1,8 @@
 package intcode;
 
+import java.math.BigInteger;
+
 public interface ReadParameter {
   String pretty();
-  int value();
+  BigInteger value();
 }
